+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update flask-smorest to version 0.31.1 to enable rapidoc OpenAPI documentation renderer
 - Default settings for sorting imports with isort (in `pyproject.toml`)
+- Add [MyST](https://myst-parser.readthedocs.io/en/latest/) markdown parser as default for documentation to support more reStructuredText features in markdown
+
+### Deprecated
+
+- Recommonmark markdown parser configuration for documentation (use the MyST parser instead)
 
 
 ## [v0.2.0]