+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update flask-smorest to version 0.31.1 to enable rapidoc OpenAPI documentation renderer
 - Default settings for sorting imports with isort (in `pyproject.toml`)
 - Add [MyST](https://myst-parser.readthedocs.io/en/latest/) markdown parser as default for documentation to support more reStructuredText features in markdown
+- Add default environment variables in `.flaskenv`
 
 ### Deprecated