+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [v0.3.0]
+
 ### Added
 
 - Updated redoc library version used in documentation to be compatible with used OpenAPI version
@@ -24,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
-- Recommonmark markdown parser configuration for documentation (use the MyST parser instead)
+- Recommonmark markdown parser configuration for documentation (use the MyST parser instead; will be removed in next release)
 
 ### Removed
 
@@ -63,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add inital flask template with documentation
 
-[unreleased]: https://github.com/buehlefs/flask-template/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/buehlefs/flask-template/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/buehlefs/flask-template/releases/tag/v0.3.0
 [v0.2.0]: https://github.com/buehlefs/flask-template/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/buehlefs/flask-template/releases/tag/v0.1.0